requirements.txt
mcp>=1.0.0
fastmcp>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0
opencv-python>=4.8.0
